move the highlighting autorun into mouse-events also
